+++ src/components/DatePicker/Month/Month.tsx
@@ -1,10 +1,10 @@
 import isBetween from 'dayjs/plugin/isBetween';
-import useKeyboardEvents from 'hooks/useKeyboardEvents';
 import { chunk, inRange } from 'lodash';
 import { useCallback } from 'react';
 import * as React from 'react';
 import dayjs, { Dayjs } from 'utils/date';
 
+import useMonthKeyboardNavigation from './hooks/useMonthKeyboardNavigation';
 import { datesWrapperStyle, weekDayStyle, weekDaysWrapperStyle } from './Month.style';
 import {
   calculatedDayIsBetween,
@@ -30,9 +30,17 @@ export type MonthProps = {
   onDaySelect?: (date: Dayjs) => void;
   selectedDays: Range;
   disabledDates?: DisabledDates;
+  isFirstCalendar?: boolean;
 };
 
-const Month: React.FC<MonthProps> = ({ year, month, onDaySelect, selectedDays, disabledDates }) => {
+const Month: React.FC<MonthProps> = ({
+  year,
+  month,
+  onDaySelect,
+  selectedDays,
+  disabledDates,
+  isFirstCalendar,
+}) => {
   const weeksWithDays = React.useMemo<WeekRow[]>(() => {
     const monthDate = currentDay.month(month).year(year).date(1);
     const daysOfMonth = monthDate.daysInMonth();
@@ -62,64 +70,12 @@ const Month: React.FC<MonthProps> = ({ year, month, onDaySelect, selectedDays, d
   const calculateIsBetween = useCallback(calculatedDayIsBetween, []);
   const calculateSelectedDayFirstOrLast = useCallback(calculateSelectedDayPosition, []);
 
-  const [focusedDay, setFocusedDay] = React.useState(0);
-
-  const calendarRef = React.useRef<HTMLTableElement>(null);
-
-  const numOfDays = currentDay.month(month).year(year).date(1).daysInMonth();
-
-  const { keyboardProps } = useKeyboardEvents({
-    events: {
-      keydown: {
-        onArrowUp: () => {
-          setFocusedDay((prevCursor) => {
-            if (prevCursor - 7 < 0) {
-              return prevCursor;
-            }
-
-            return prevCursor - 7;
-          });
-        },
-        onArrowDown: () => {
-          setFocusedDay((prevCursor) => {
-            if (prevCursor + 7 > numOfDays) {
-              return prevCursor;
-            }
-
-            return prevCursor + 7;
-          });
-        },
-        onArrowMove: (__, direction) => {
-          if (direction === 'left') {
-            setFocusedDay((prevCursor) => {
-              if (prevCursor === 0) {
-                return 0;
-              }
-
-              return prevCursor - 1;
-            });
-          }
-
-          if (direction === 'right') {
-            setFocusedDay((prevCursor) => {
-              if (prevCursor === numOfDays) {
-                return numOfDays;
-              }
-
-              return prevCursor + 1;
-            });
-          }
-        },
-      },
-    },
+  const { focusedDay, setFocusedDay, calendarRef, keyboardProps } = useMonthKeyboardNavigation({
+    selectedDays,
+    isFirstCalendar,
+    month,
+    year,
   });
-
-  React.useEffect(() => {
-    if (calendarRef.current) {
-      const selectCursor = calendarRef.current.querySelector(`div[tabindex = "0"]`);
-      (selectCursor as HTMLInputElement)?.focus();
-    }
-  }, [focusedDay]);
 
   return (
     <React.Fragment>
